+++ packages/studio-website/src/pages/importing/index.tsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React, { useState } from 'react';
 import ImportApp, { ISchemaOptions, transMappingSchemaToOptions } from '@graphscope/studio-importor';
 import { queryPrimitiveTypes, uploadFile, getSchema, getDatasourceById } from './services';
 import { useContext } from '../../layouts/useContext';
@@ -12,11 +12,14 @@ import localforage from 'localforage';
 import { FormattedMessage } from 'react-intl';
 import DownloadLoadConfig from './download-load-config';
 import FeatureCase from '../../components/feature-case';
+import DataBind from './data-bind';
+import DataLoad from './data-load';
 interface ISchemaPageProps {}
 const { GS_ENGINE_TYPE } = window;
 const SchemaPage: React.FunctionComponent<ISchemaPageProps> = props => {
   const { store, id } = useContext();
   const { graphId, draftId } = store;
+  const [refreshIndex, setRefreshIndex] = useState(1);
 
   const queryBoundSchema = async (): Promise<ISchemaOptions> => {
     if (graphId === draftId) {
@@ -52,6 +55,9 @@ const SchemaPage: React.FunctionComponent<ISchemaPageProps> = props => {
     const localFiles = getLocalFiles();
     return localFiles;
   };
+  const refresh = () => {
+    setRefreshIndex(prev => prev + 1);
+  };
 
   return (
     <Section
@@ -64,6 +70,7 @@ const SchemaPage: React.FunctionComponent<ISchemaPageProps> = props => {
     >
       <EmptyModelCase />
       <ImportApp
+        refreshIndex={refreshIndex}
         key={graphId}
         appMode="DATA_IMPORTING"
         queryBoundSchema={queryBoundSchema}
@@ -81,9 +88,12 @@ const SchemaPage: React.FunctionComponent<ISchemaPageProps> = props => {
       >
         <Toolbar style={{ top: '12px', left: '24px', right: 'unset' }} direction="horizontal">
           <SelectGraph id={id} />
-          <StartImporting id={id} />
-          <FeatureCase match="DOWNLOAD_DATA_TASK_CONFIG">
-            <DownloadLoadConfig id={id} graphId={graphId} />
+          <FeatureCase match="SCHEMA_UPDATE">
+            <DataBind id={id} refresh={refresh} />
+            <DataLoad id={id} />
+          </FeatureCase>
+          <FeatureCase match="BATCH_LOAD_DATA">
+            <StartImporting id={id} />
           </FeatureCase>
         </Toolbar>
       </ImportApp>