+++ packages/studio-explore/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphscope/studio-explore
 
+## 0.1.16
+
+### Patch Changes
+
+- update website
+- Updated dependencies
+  - @graphscope/studio-components@0.1.18
+  - @graphscope/studio-importor@0.1.18
+  - @graphscope/studio-query@0.1.18
+  - @graphscope/studio-driver@0.1.16
+  - @graphscope/studio-graph@0.1.18
+
 ## 0.1.15
 
 ### Patch Changes