+++ packages/studio-graph/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphscope/studio-graph
 
+## 0.1.18
+
+### Patch Changes
+
+- update website
+- Updated dependencies
+  - @graphscope/studio-components@0.1.18
+  - @graphscope/use-zustand@0.1.10
+
 ## 0.1.17
 
 ### Patch Changes