+++ packages/studio-importor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @graphscope/studio-importor
 
+## 0.1.18
+
+### Patch Changes
+
+- update website
+- Updated dependencies
+  - @graphscope/studio-components@0.1.18
+  - @graphscope/studio-server@1.0.15
+  - @graphscope/use-zustand@0.1.10
+
 ## 0.1.17
 
 ### Patch Changes