+++ packages/studio-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphscope/studio-server
 
+## 1.0.15
+
+### Patch Changes
+
+- update website
+
 ## 1.0.14
 
 ### Patch Changes